Suppress dead code warning for DfuState::read_from_device()
Although we don't currently call this, instead preferring
DfuStatus::read_from_device() which gives us strictly more information,
it serves to document a piece of the protocol that isn't used anywhere
else. As such, keep it around so that this library is a complete
description of all pieces of the protocol we know about.