protocol: Validate that the device returns enough data
Currently, we don't check the return value of get_feature_report(),
meaning if we get an unexpected short read we might treat some of the
initial zeroes in the buffer as real data from the device. For uploads,
the situation is worse, as we might accidentally look at data from the
previous read. Fix both these issues by bailing out on a short read and
zeroing the upload buffer every loop iteration.