Overhaul device matching, warn on untested devices
This is a significant refactoring of the logic that determines which
device we should perform an operation on. In addition to moving the list
of known devices and associated logic into a new `bose_dfu::device_ids`
module, this commit introduces the concept of "device risks", which when
present trigger additional warnings and require a new `-f` command-line
flag to bypass.