Call ensure_idle() during leave-dfu
I'm not sure if Bose's custom leave DFU request requires the device to
be in dfuIDLE like other requests do, but adding this call has the very
useful side effect of logging what error state, if any, the device was
in. This improves user experience because any error states that prevent
leave-dfu from succeeding (such as errFIRMWARE) are now printed when it
runs.